fix: federation subgraph SDL should have federation directives
We're doing validation on SDL printed from the registry in `grafbase dev`, but
this wasn't printing the federation subgraph fields & directives, even when
subgraph support was enabled.

This is problematic because a subgraph that defines no query fields is valid,
but its schema won't be valid if you don't add the federation plumbing to it.