chore: give the registry a haircut (#1625)
I've been profiling with CCs actual queries & schema today, and the
number one consumer of memory with these tiny queries is the registry.
As always when I go on one of these missions, there's a ton of low
hanging fruit we can get rid of. MetaField in particular was _huge_ - I
forget the numbers, but 700 bytes or so per field.

In my profiling this reduces the total memory use from ~16.5mb to ~11.5.
This is all on arm rather than wasm so actual numbers probably vary, but
it's an improvement either way.