feat: start using registry-v{1,2} crates
Updates engine to use registry-v1, and the parsers to (mostly) use registry-v1
rather than the engine.

Sorry for the size of this one.  I could split it up but I'm not sure if it's
especially worth it.

This work is not finished by this PR, but it's large enough as it is.  I'll
make some separate PRs on top of this one to finish it before I merge to main.