feat: add registry-v2 crate
As I mentioned in the meeting this morning, I've been working on refactoring
the registry to a more compact & performant representation (hopefully also
easier to use, but time will tell).

I've done this in a new crate `registry-v2`.  Separated from the `engine`
because having the registry inside engine always limited how modular we could
be: if the engine needed to call code that needed the registry that code
usually to live inside the `engine` crate: leading to a severely bloated crate
that takes an age to compile (waiting 30 seconds for feedback on my changes
while working on this was not fun)

The structure of this crate follows the same pattern I've been using in
cynic-parser (which is itself similar to patterns used in engine-v2 &
elsewhere).  Each data type:

- Is stored as a "record" in a flat array in the Regsitry struct.
- Has an ID that is just a thin wrapper around index into the array.
- Has reader structs built on top of IDs that provide a reasonably ergonomic
  API for reading the data (known as walkers in some of our other
  implementations of this pattern).  Since these are the primary way to
  interact with the data, they just take the name of the data, no
  reader/walker postfix.
- Functions on a shared Writer struct that can be used to populate the
  Registry.

Writing code for all of the above is quite laborious (but not that complicated)
so I've borrowed a generator that I wrote for cynic-parser to do a lot of it.
All the code in the `generated` folder was produced by this.  The generator
will be in a separate PR.  Note that not everything is worth generating - any
special cases that are easier to write by hand should be.

Some of the code in this crate is just copied & pasted from the old registry.
We could update that code to use the above pattern sometime, but this change is
already massive.

Any questions, comments, criticisms - give them to me.