cli: implement minimum support for NixOS
Currently, if you start `grafbase dev` or any command that reads the configuration
with configuration in a grafbase.config.ts file, on NixOS, you will see this:

```
Grafbase CLI 0.69.1

Error: could not load grafbase/grafbase.config.ts
Caused by: Could not start dynamically linked executable: /home/tom/.grafbase/bun/bun
NixOS cannot run dynamically linked executables intended for generic
linux environments out of the box. For more information, see:
https://nix.dev/permalink/stub-ld
```

This is because the CLI downloads a bun binary from GitHub releases. That
will not work on NixOS.

This commit implements the minimum version of NixOS support: use the `bun`
executable from `$PATH`, and if that is missing, fail with a helpful error.

Ideally, we would want to run the exact version of bun the CLI is tested
with, but we do not want to run `nix` on behalf of the user. We could point
them to the right command, or even better, start enforcing that the derivation
for the CLI in this repo is working in CI. In the meantime, this commit
is an improvement on the status quo.