sdk: add ability to pass GraphQL SDL type definitions to standalone graphs
This lets users mix GraphQL SDL and SDK to their preference when defining
their configuration. The idea and implementation are both very basic.

The API takes the shape of a new argument to `graph.Standalone()`:

```typescript
const g = graph.Standalone({
  typeDefs: /* GraphQL */ `
    type Invoice {
      id: ID!
      invoiceNumber: String!
      dueDate: Date!
      totalAmount: Int!
    }

    extend type Query {
      invoiceByNumber(invoiceNumber: String!): Invoice @resolver(name: "invoice/byNumber")
    }
  `
})

```

The implementation is equally straightforward: we append the GraphQL
that was passed in at the end of the generated SDL configuration.

Potential issues:

- Ordering: we have a set ordering for schema items in the SDK. Appending
  the string as-is will not respect it.
- Error message: it will not be clear where the error originated, but this
 is no worse than if you defined your whole config in SDL in the first place.

The issues seem to be minor enough to overlook, since it's a small change
that enables new use cases.

closes GB-6547