typed-resolvers: implement correct codegen for interfaces (#1648)
Before this commit, codegen would generate a TS interface with the
fields of the GraphQL interface. But we are generating code for
resolvers, and the resolver should not return only the interface fields,
but a whole object that implements the interface, including
`__typename`. This commit changes the generated code to a TypeScript
union of all implementers of the interface.

closes GB-5171