Stabilize resolver type code generation (codegen experimental feature) (#1655)
This stabilization PR does two things:

- Introduce the internal `@codegen` directive to configure codegen,
replacing `@experimental(codegen: true)`. The experimental flag
continues working for now but it is marked as deprecated in the SDK, so
it will be a visible warning in users' IDE.
- Make the path where the generated code is written configurable.

We will need a SDK release after this is merged to expose the new
configuration. In practice, it looks like this:

```ts
export default config({
  graph: g,
  codegen: {
     enabled: true,
     path: 'cat/meow'
  }
});
```

closes GB-6468
closes GB-5246
closes GB-5245
closes GB-5244