feat: code generator for registry-v2 (#1654)
As outlined in [the PR that adds registy-v2][1] a lot of that crate is
generated. This PR contains the generator. It:

- Reads the data structure from a GraphQL file (SDL for life).
- Generates the records, readers and neccesary plumbing for everything
to work.
- Manually implements some things that would usually be derived (i.e.
`Debug` & `PartialEq`), but can't because of the patterns being used.
- Adds some serde annotations on things to make our JSON smaller. We've
used serde_as for this in the past, but why have a macro do work on
every compile when you can have a generator do it once (with the bonus
that it's more obvious what transforms are taking place)

This crate started as a c&p from cynic-parser and is very quick & dirty.
it does what it needs to do and nothing more, it probably has bugs or
strange behaviour if you try to do new things with it. But it's worked
so far, and is far easier than writing all this by hand (and makes
changing things en masse relatively easy).

[1]: https://github.com/grafbase/grafbase/pull/1651