chore: extract registry-v1 crate (#1652)
This PR adds a `registry-v1` crate. This is basically a copy & paste of
the registry struct and its component structs & enums into a separate
crate.

I've done this for a similar reason to adding `registry-v2` as a
separate crate: having everything live in engine makes it hard to split
out individual modules into crates.

At runtime the engine is going to use registry-v2. But I've not yet
updated any of the parsers to use registry-v2. They all still generate a
v1 Registry, which we then upgrade to registry-v2 before feeding to the
engine.

Since the engine sometimes needs that conversion code (for tests mostly)
and that conversion code needs access to both v1 & v2, we'd need to
build the conversion inside engine-v1 if we hadn't split this out. Which
would further tie us into having `engine` in the dep tree for
everything. Having it separate means we can have the parsers just depend
on `registry-v{1,2}` and ditch their engine dependency (eventually at
least)

We may want to update the parsers to drop registry-v1 sometime, but I'm
not sure if we ever want them to be writing a registry-v2 directly. It's
got some nice properties for the engine runtime, but it has requirements
that maybe aren't great for building up a schema. We can cross that
bridge when & if we come to it though.