engine-v2: implement and test empty schema
We want to clean up the empty state for federated graphs in the API. This
is the state right after the federated graph is created, when no subgraph
has been published yet. This commit implements an "empty" state for FederatedGraph
that cleanly converts to an engine_v2::Schema, with a test to check that
the engine can start with that schema.

Part of GB-6537