fix: correctly prune the caching registry (#1677)
The caching registry was just stripping out types that didn't have a
cache_config on them. This was too simple and had some problems:

- If a field on a cached type contains e.g. `String!` then the registry
is invalid as `String!` isn't a cached type and isn't included in the
registry.
- If a cached type is nested inside a non-cached type then the path to
that cached type won't be in the `Registry` so the visitor in
`validation` won't have `current_type` info and that type won't be
correctly included in the cache key in the gateway.
- Possibly other problems, but those are the main ones.

This PR fixes that: we now do a proper traversal of the type graph to
make sure we include all the cached types, their cached fields, the
types their cached fields point to and any intermediate types between
the cached type and the root.

For some schemas this might lead to a lot more stuff being included in
the registry, which might have impacts on loading times etc. Hopefully
nobody is caching any seriously nested types, but we might need to keep
an eye on this.

There's also some unrelated snapshot changes in here, pay them no mind.