sdk: better error reporting for the empty object case
Before this commit, we show a very vague and unhelpful error from
async-graphql-parser. This commit moves the error earlier on in the
process when rendering SDL from SDK.

The potentially controversial bit in this PR is that I think we don't
throw in any other case in that stage of the process. But better handling
on the engine side is not possible without replacing async-graphql-parser.

closes GB-6604