chore: add partial caching to experimental directive (#1729)
Want to use this as a feature flag while I develop so I don't have to
worry about breaking existing users. It's not even close to ready for
public consumption yet so don't want to add it to the SDK yet (although
I might need to at some point).

I've added a warning when this is enabled just incase users stumble
across it.

Also removed some AI & KV stuff, updating them to put out warnings if a
user is still using them (iirc they're deprecated in the SDK, but this
is another signal that doesn't hurt). Also put a warning on codegen to
tell users it's not experimental.