feat: partial cache fetching logic (#1735)
This implements the logic for building cache keys out of the various
cachable query partitions, along with fetching them from the cache and
then determining the final query to send along to the executor.

I've introduce the notion of "phases" to the partial-caching crate - we
enter a "CacheFetchPhase", represented by a struct, that exposes
functions to handle all the logic of this phase. When done you can call
".finish()" which will either move to an execution phase or some as yet
unnamed final phase.

I've also started integrating this with the gateway_core crate. This is
a very broken implementation, but one that will only be called if you
enable partial caching in your registry (don't do that) and you have the
cargo feature flag enabled (which it is in the CLI, so that I can work
on this easier, but won't be in production till this is closer to
ready).

The cache key calculations in this differ in a couple of ways from the
existing setup:

1. I'm using blake3 for the hashing instead of SipHash. I didn't feel
comfortable using a non-cryptographic hash function - from a security
standpoint, and also because collisions will mess this partial caching
thing up horrendously, and I do not fancy debugging that down the line.
2. I'm hashing the JSON serialised form of the data, instead of doing
order agnostic things with Objects like we do in the existing hash
functions. This is a bit annoying, although I'm not sure how much of an
issue this is in practice. It was way faster to implement though. I've
raised a linear issue to revisit this if I get time/we decide that it
matters.

Fixes GB-6741