composition: ignore subgraph federation mandated fields and types (#1743)
Subgraphs must expose the _entities and _service fields (and the
corresponding _Entity and _Service types) in order to function as
subgraphs behind the federated gateway. Depending on how you introspect
them, these fields and types may end up in the subgraph SDL sent for
composition. Since these fields are not part of the domain API, and they
do not compose cleanly (duplicated fields and types across subgraphs),
composition can and should ignore them.

This is what is implemented in this commit.

closes GB-6830