feat: tracing tweaks for CLI (#1739)
I was trying to debug partial caching yesterday but my `trace::info!`
calls weren't appearing in my terminal, causing much confusion. This
turned out to be because my trace calls were in `gateway_core` which is
not in the log filter that's set when you do `--trace 2`. So, I've done
a few things to fix:

1. Added a default of info to the `--trace 2` filter, to make it easier
to get everything. We have too many crates (and introduce new ones all
the time) to expect this list to always be exhaustive.
2. I've added a hidden `--custom-trace-filter` argument that lets you
provide a specific filter on the CLI.
3. I've added support for the `RUST_LOG` env var if neither of the CLI
arguments are provided.

This made it much easier to debug my code, as I could ask for
`gateway_core` traces specifically, and avoid the noise that all the
other log filter flags introduce. It was _so much easier_ from that
point on.

Some questions for the team though:

1. `--trace` is a bit weird in that it takes a number, where 1 doesn't
do anything and anything up from 2 does the same thing. Should we
rethink this? Either ditch the number or make 1 do something?
2. `--custom-trace-filter` is hidden, because I didn't feel like
changing public interfaces. But should we make it shown? Or should I
just remove it in favour of always using the env var?
3. I debated using a grafbase specific env var like `GRAFBASE_LOG`
instead of `RUST_LOG`. Thoughts?

Also welcome any other thoughts anyone has.