feat: partial cache settings propagation (#1746)
This updates the partial caching implementation to handle propagate
settings from types & settings onto their descendants. So if a type
`User` as a cache control, then all the fields inside `User` will
inherit those settings.

If a field _and_ the type that field contains both have a cache control
then we'll give preference to the field setting, as it's more specific.

This complicated the fragment logic a bit, as we have to propagate
through a fragment spread as well - which means we potentially need to
process fragments multiple times if they are spread multiple times with
different inherited cache settings. So were we were before just keying
fragments by their id, we're now using a `FragmentKey` - which takes the
inhereted cache control value into account as well.

Fixes GB-6777