nix: restore cli-app and cli derivations
Since c896aa3a842119d476936bb8665b5a35792c9bfb, we haven't been able to
build the Grafbase CLI as a nix derivation, due to the breaking upgrade
to pnpm 9 and the lack of support for pnpm 9 in pnpm2nix. Fortunately,
https://github.com/NixOS/nixpkgs/pull/290715 was recently merging into
nixpkgs, so we can use pnpm_9.fetchDeps to re-instate the cli-app
derivation, and as a cascading effect build the cli derivation again.