postgres connector: multiple relations between same tables
Given a table with two foreign keys referencing the same table, for example:

```sql
CREATE TABLE public.colors (
    id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
    name VARCHAR(255) NOT NULL UNIQUE,
    rgb INT NOT NULL
);

CREATE TABLE public.users (
    id INT PRIMARY KEY GENERATED ALWAYS AS IDENTITY,
    name VARCHAR(255) NOT NULL UNIQUE,
    favorite_color_id INT NOT NULL REFERENCES colors(id),
    least_favorite_color_id INT REFERENCES colors(id)
);
```

Previously, we would have only a `colors` field on the `users` type, using
one of the two relations at random.

This commit changes the connector so we have two fields:
`colorsByFavoriteColorId` and `colorsByLeastFavoriteColorId`. The fields
on `colors` that walk the relation in the other direction are also adapted.

closes GB-6878