feat: calculate output shapes in partial caching (#1786)
I'm adding support for `@defer` to our partial caching. When handling
responses & cache entries in the presence of `@defer` we need to be able
to tell whether or not the field we're currently processing is inside a
`@defer` or not.

This sounds simple but has a lot of complicating factors:

1. The JSON structure of responses & cache chunks doesn't always match
the GraphQL AST structure. Fragments add an additional level of nested
to the AST, but not to the JSON.
2. Type conditions on fragments add branching into the equation.
3. A given field of an object could theoretically be present both inside
and outside a defer, with the difference between those selections being
nested several levels deep in the GraphQL AST.

At present most of the partial caching response handling is a fairly
simple deep merge over the JSON of the response & the cache entry. But
handling these extra concerns while doing that sounds extremely
difficult, and there's likely going to be a few different merge
implementations, so we might even need to handle it in several places.

This commit adds my solution for this: we precaclulate the output shape
of the individual query partitions. This:

1. Gives us a heirarchy with the same structure as the JSON output, for
easy traversal when doing a merge.
2. Should make handling fragments easier - we just need to read the
`__typename` and select a shape based on that.
3. Handles propagating the defer label to fields that are exclusively in
the defer.

Which should make implementing the rest of the defer stuff easier.
Hopefully... 🤞

Fixes GB-6747