feat: merge partial cache entries into initial response (#1799)
#1791 added an `OutputStore` and some code to write the initial response
from the engine into that. This PR handles merge cache response into the
OutputStore. This is a bit more involved than the initial response,
because we need to be careful not to add any deferred fields into the
response, and instead save them for when the deferred chunks come in.

Still a bunch of TODOs in here.  Will revisit these later though.

Fixes GB-6940