feat: incremental response merging in partial-caching
With the code I've written so far in partial-caching I'm able to take the
initial response and put that into `OutputStore` and take the non-deferred
cache chunks and put them into `OutputStore`.

This PR adds the two final pieces before I start wiring things up:

1. Updated the cache chunk merging so that it can include deferred chunks as
well as skipping them.  This means that after the initial response has been
written to users we can just merge the deferred chunks into the overall
response, ready for the incremental payloads.
2. Wrote the code to merge in deferred payloads as well.

As always with this partial caching, I have absolutely littered the code with
todos, which I will come back to tackle later.