chore: split engine-response crate out from engine (#1801)
I've been trying to keep the dependency tree for the `partial-caching`
crate fairly small. But it looks like I might be about to need the
`IncrementalResponse` struct in order to handle deferred payloads.

I am going to try and avoid this, but just incase - this PR splits an
`engine-response` crate out from `engine`. There should be no functional
changes in this, just copying these things:

1. `engine::{Response, IncrementalResponse}`
2. `ServerError` since that is part of `engine::Response`.
3. I added a new `CacheControl` struct to this engine-response crate to
avoid pulling in `registry-v2`. This doesn't have all of the fields that
the `registry-v2` one does, but it looked like they weren't all needed
for the value in the `Response`. If I've missed a field we can add it
again later.
4. Deleted the `query-path` module from engine. I'd already extracted
this to its own crate, just apparently forgot to delete the file from
engine.

Part of GB-6941