fix(engine-v2): Proper error propagation of authenticated & requiresScopes
Until now we were entirely failing the query if those conditions weren't
valid. Now it only invalidates the right part of the query and planning
can still be cached.

When building the operation we keep track all of the conditions that may
change the query. Those are only executed when building the "output
shape". If access is denied we keep the generated error and will add it
to the responses whenever the field would be accessed.